Broke model up into files
