+++ view.rb
@@ -1,0 +1,5 @@
+class View
+
+
+
+end
